refactor: extract app layout into default layout
Move UHeader, UMain and UFooter from app.vue to layouts/default.vue.
Introduce NuxtLayout so different pages can use different layouts.
